+++ service/src/java/org/apache/hive/service/servlet/HS2LeadershipStatus.java
@@ -27,7 +27,6 @@ import javax.servlet.http.HttpServletResponse;
 
 import org.apache.hive.http.HttpConstants;
 import org.apache.hive.http.HttpServer;
-import org.apache.hive.service.server.HiveServer2;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -35,9 +34,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 
 /**
  * Returns "true" if this HS2 instance is leader else "false".
- * Invoking a "DELETE" method on this endpoint will trigger a failover if this instance is a leader.
  * hadoop.security.instrumentation.requires.admin should be set to true and current user has to be in admin ACLS
- * for accessing any of these endpoints.
+ * for accessing this endpoint.
  */
 public class HS2LeadershipStatus extends HttpServlet {
   private static final Logger LOG = LoggerFactory.getLogger(HS2LeadershipStatus.class);
@@ -65,82 +63,10 @@ public class HS2LeadershipStatus extends HttpServlet {
     response.flushBuffer();
   }
 
-  private void setResponseHeaders(final HttpServletResponse response) {
+  void setResponseHeaders(final HttpServletResponse response) {
     response.setContentType(HttpConstants.CONTENT_TYPE_JSON);
     response.setHeader(HttpConstants.ACCESS_CONTROL_ALLOW_METHODS,
-      HttpConstants.METHOD_GET + "," + HttpConstants.METHOD_DELETE);
+        HttpConstants.METHOD_GET + "," + HttpConstants.METHOD_DELETE);
     response.setHeader(HttpConstants.ACCESS_CONTROL_ALLOW_ORIGIN, HttpConstants.WILDCARD);
-  }
-
-  private class FailoverResponse {
-    private boolean success;
-    private String message;
-
-    FailoverResponse() {
-    }
-
-    public boolean isSuccess() {
-      return success;
-    }
-
-    public void setSuccess(final boolean success) {
-      this.success = success;
-    }
-
-    public String getMessage() {
-      return message;
-    }
-
-    public void setMessage(final String message) {
-      this.message = message;
-    }
-  }
-
-  @Override
-  public void doDelete(final HttpServletRequest request, final HttpServletResponse response) throws IOException {
-    // strict admin check -
-    // allows ONLY if hadoop.security.instrumentation.requires.admin is set to true
-    // when hadoop.security.instrumentation.requires.admin is set to true, checks if hadoop.security.authorization
-    // is true and if the logged in user (via PAM or SPNEGO + kerberos) is in hive.users.in.admin.role list
-    final ServletContext context = getServletContext();
-    if (!HttpServer.isInstrumentationAccessAllowedStrict(context, request, response)) {
-      LOG.warn("Unauthorized to perform DELETE action. remoteUser: {}", request.getRemoteUser());
-      return;
-    }
-
-    setResponseHeaders(response);
-
-    LOG.info("DELETE handler invoked for failover..");
-    ObjectMapper mapper = new ObjectMapper();
-    FailoverResponse failoverResponse = new FailoverResponse();
-    AtomicBoolean isLeader = (AtomicBoolean) context.getAttribute("hs2.isLeader");
-    if (!isLeader.get()) {
-      String msg = "Cannot failover an instance that is not a leader";
-      LOG.info(msg);
-      failoverResponse.setSuccess(false);
-      failoverResponse.setMessage(msg);
-      mapper.writerWithDefaultPrettyPrinter().writeValue(response.getWriter(), failoverResponse);
-      response.setStatus(HttpServletResponse.SC_FORBIDDEN);
-      return;
-    }
-
-    HiveServer2.FailoverHandlerCallback failoverHandler = (HiveServer2.FailoverHandlerCallback) context
-      .getAttribute("hs2.failover.callback");
-    try {
-      String msg = "Failover successful!";
-      LOG.info(msg);
-      failoverHandler.failover();
-      failoverResponse.setSuccess(true);
-      failoverResponse.setMessage(msg);
-      mapper.writerWithDefaultPrettyPrinter().writeValue(response.getWriter(), failoverResponse);
-      response.setStatus(HttpServletResponse.SC_OK);
-    } catch (Exception e) {
-      String errMsg = "Cannot perform failover of HS2 instance. err: " + e.getMessage();
-      LOG.error(errMsg, e);
-      failoverResponse.setSuccess(false);
-      failoverResponse.setMessage(errMsg);
-      mapper.writerWithDefaultPrettyPrinter().writeValue(response.getWriter(), failoverResponse);
-      response.setStatus(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
-    }
   }
 }
